core/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nefertem-core"
version = "2.0.4"
description = "Python data validation library"
readme = "README.md"
authors = [
    { name = "Fondazione Bruno Kessler", email = "dslab@fbk.eu" },
    { name = "Matteo Martini", email = "mmartini@fbk.eu" }
]
license = { file = "LICENSE.txt" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["data", "validation", "quality"]
dependencies = [
    "boto3>=1.28.0, <1.29.0",
    "aiobotocore>=2.5.0, <2.8",
    "pydantic~=1.10, >=1.10.8",
    "pandas>=1.2, <3",
    "sqlalchemy~=1.4",
    "psutil>=5.9",
    "requests~=2.31",
    "adbc-driver-manager==0.7.0",
    "polars",
]

requires-python = ">=3.9"

[tool.setuptools.packages.find]
exclude = ["docs*", "tests*", "plugins*", "operations*"]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "moto",
    "bumpver",
]
docs = [
    "Sphinx==5.3.0",
    "pydata-sphinx-theme==0.13.3",
    "autodoc-pydantic==1.8.0",
    "numpydoc==1.5.0"
]

[project.urls]
Homepage = "https://github.com/scc-digitalhub/nefertem"

[tool.flake8]
max-line-length = 120

[tool.ruff]
line-length = 120

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.bumpver]
current_version = "2.0.3"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = false
tag             = false
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
